requirements.txt

Flask==3.0.2
requests==2.31.0
pandas==2.2.2
gunicorn==21.2.0
python-dotenv==1.0.1
